requirements_dev.txt
flake8
transifex-client
tox
django-nose
mkdocs
ghp-import
model_mommy
coverage==3.7.1
django-coverage
coveralls
